cast to int

+++ ksql-test-util/src/main/java/io/confluent/ksql/test/util/KafkaEmbedded.java
@@ -269,8 +269,8 @@ class KafkaEmbedded {
     effectiveConfig.put(KafkaConfig.AutoCreateTopicsEnableProp(), true);
     effectiveConfig.put(KafkaConfig.MessageMaxBytesProp(), 1_000_000);
     effectiveConfig.put(KafkaConfig.ControlledShutdownEnableProp(), true);
-    effectiveConfig.put(KafkaConfig.ZkSessionTimeoutMsProp(), ZK_SESSION_TIMEOUT.toMillis());
-    effectiveConfig.put(KafkaConfig.ZkConnectionTimeoutMsProp(), ZK_CONNECT_TIMEOUT.toMillis());
+    effectiveConfig.put(KafkaConfig.ZkSessionTimeoutMsProp(), (int) ZK_SESSION_TIMEOUT.toMillis());
+    effectiveConfig.put(KafkaConfig.ZkConnectionTimeoutMsProp(), (int) ZK_CONNECT_TIMEOUT.toMillis());
 
     effectiveConfig.putAll(initialConfig);
     effectiveConfig.put(KafkaConfig.LogDirProp(), logDir.getAbsolutePath());
